requirements.txt
fastapi
uvicorn[standard]
pypdf
python-dotenv
